MNES-1238: refactor - renaming
