+++ src/app/[locale]/viewer/_components/submodel-elements/document-component/DocumentComponent.tsx
@@ -18,7 +18,7 @@ import { isValidUrl } from 'lib/util/UrlUtil';
 import { encodeBase64 } from 'lib/util/Base64Util';
 import { useAsyncEffect } from 'lib/hooks/UseAsyncEffect';
 import Image from 'next/image';
-import { getAttachmentFromSubmodelElement } from 'lib/services/multiple-repository-access/MultipleRepositorySearchActions';
+import { getAttachmentFromSubmodelElement } from 'lib/services/repository-access/RepositorySearchActions';
 
 enum DocumentSpecificSemanticId {
     DocumentVersion = 'https://admin-shell.io/vdi/2770/1/0/DocumentVersion',
