+++ src/app/[locale]/viewer/discovery/_components/DiscoveryListView.tsx
@@ -15,8 +15,8 @@ import { useEnv } from 'app/env/provider';
 import { encodeBase64 } from 'lib/util/Base64Util';
 import ListHeader from 'components/basics/ListHeader';
 import { performDiscoveryAasSearch, performRegistryAasSearch } from 'lib/services/search-actions/searchActions';
-import { performSearchAasFromAllRepositories } from 'lib/services/multiple-repository-access/MultipleRepositorySearchActions';
-import { RepoSearchResult } from 'lib/services/multiple-repository-access/MultipleRepositorySearchService';
+import { performSearchAasFromAllRepositories } from 'lib/services/repository-access/RepositorySearchActions';
+import { RepoSearchResult } from 'lib/services/repository-access/RepositorySearchService';
 
 export const DiscoveryListView = () => {
     const [isLoadingList, setIsLoadingList] = useState(false);
