+++ src/lib/services/repository-access/PrismaConnector.ts
@@ -1,7 +1,7 @@
 import { prisma } from 'lib/database/prisma';
 import { ConnectionType } from '@prisma/client';
-import { IPrismaConnector } from 'lib/services/multiple-repository-access/PrismaConnectorInterface';
-import { PrismaConnectorInMemory } from 'lib/services/multiple-repository-access/PrismaConnectorInMemory';
+import { IPrismaConnector } from 'lib/services/repository-access/PrismaConnectorInterface';
+import { PrismaConnectorInMemory } from 'lib/services/repository-access/PrismaConnectorInMemory';
 
 export type DataSourceFormData = {
     id: string;
