+++ src/lib/services/repository-access/PrismaConnectorInterface.ts
@@ -1,5 +1,5 @@
 import { ConnectionType } from '@prisma/client';
-import { DataSourceFormData } from 'lib/services/multiple-repository-access/PrismaConnector';
+import { DataSourceFormData } from 'lib/services/repository-access/PrismaConnector';
 
 export interface IPrismaConnector {
     getConnectionData(): unknown;
